+++ pages/todo-components/index.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Link from 'next/link';
+import { RootStyles } from '../../styles/common';
+
+export default () => (
+  <>
+    <RootStyles>
+      <span>Components List</span>
+    </RootStyles>
+  </>
+);
